Created private implementation for TarHeader
