+++ tar/lib/src/constants.dart
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import 'package:tar/src/format.dart';
+import 'format.dart';
 
 /// Magic values to help us identify the TAR header type.
 const magicGNU = 'ustar ';
